// src/types/quest.types.ts
export type ActiveQuestStatus = 'Running' | 'Done' | 'Error';
export interface ActiveQuest {
  id: string;
  name: string;
  reward: string;
  remaining: number;
  status: ActiveQuestStatus;
}
export interface UserInfo {
  username: string;
  id: string;
}